Rename stream to device
