+++ device/device_test.go
@@ -1,4 +1,4 @@
-package stream
+package device
 
 import (
 	"bytes"
