+++ device/doc.go
@@ -1,2 +1,2 @@
 // stream device package provides a way to create and manage virtual lab devices that generate and consume data streams
-package stream
+package device
